3D Model Updates and ButtonScript
Fixed rotation on player snake so snake faces right direction when moving, removed my test script for button "animation" and added everything to the actual button script, added new suggestion for floor tiles
