--- SoTA - PreProd/Assets/Scripts/Karin_ButtonScriptTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f3dfb68aa6242274ab794c797d9fbca9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
